deprecate .iterate() method
