+++ examples/ubxpoller.py
@@ -1,133 +1,156 @@
 """
-ubxpoller.py
+ubxpoller2.py
 
-This example illustrates a simple implementation of a
-'pseudo-concurrent' threaded UBXMessage configuration
-polling utility.
+This example illustrates a permutation of 'pseudo-concurrent'
+threaded read and write UBX message processing using
+queues to pass messages between threads.
 
-(NB: Since Python implements a Global Interpreter Lock (GIL),
-threads are not truly concurrent.)
+Press CRTL-C to terminate.
 
-It connects to the receiver's serial port and sets up a
-UBXReader read thread. With the read thread running
-in the background, it sends a variety of CFG POLL
-messages to the device. The read thread reads and parses
-any responses to these polls and outputs them to the terminal.
+NB: Since Python implements a Global Interpreter Lock (GIL),
+threads are not truly concurrent. True concurrency could be
+achieved using multiprocessing (i.e. separate interpreter
+processes rather than threads) but this is non-trivial in
+this context as serial streams cannot be shared between
+processes. A discrete hardware I/O process must be implemented
+e.g. using RPC server techniques.
 
-The response may be an ACK-ACK acknowledgement message followed
-by the poll response itself, or an ACK-NAK message signifying
-that this particular configuration message type is not supported
-by the receiver.
+Created on 07 Aug 2021
 
-Created on 2 Oct 2020
-
-@author: semuadmin
+:author: semuadmin
+:copyright: SEMU Consulting © 2021
+:license: BSD 3-Clause
 """
 # pylint: disable=invalid-name
 
 from sys import platform
-from io import BufferedReader
-from threading import Thread, Lock
+from threading import Thread, Event, Lock
+from queue import Queue
 from time import sleep
 from serial import Serial
-from pyubx2 import (
-    UBXMessage,
-    UBXReader,
-    POLL,
-    UBX_MSGIDS,
-)
-
-# initialise global variables
-reading = False
+from pyubx2 import UBXReader, UBXMessage, POLL, UBX_PROTOCOL
 
 
-def read_messages(stream, lock, ubxreader):
+def read_data(
+    stream: object,
+    ubr: UBXReader,
+    queue: Queue,
+    lock: Lock,
+    stop: Event,
+):
     """
-    Reads, parses and prints out incoming UBX messages
+    Read and parse incoming UBX data and place
+    raw and parsed data on queue
     """
     # pylint: disable=unused-variable, broad-except
 
-    while reading:
+    while not stop.is_set():
         if stream.in_waiting:
             try:
                 lock.acquire()
-                (raw_data, parsed_data) = ubxreader.read()
+                (raw_data, parsed_data) = ubr.read()
                 lock.release()
                 if parsed_data:
-                    print(parsed_data)
+                    queue.put((raw_data, parsed_data))
             except Exception as err:
                 print(f"\n\nSomething went wrong {err}\n\n")
                 continue
 
 
-def start_thread(stream, lock, ubxreader):
+def write_data(stream: object, queue: Queue, lock: Lock, stop: Event):
     """
-    Start read thread
-    """
-
-    thr = Thread(target=read_messages, args=(stream, lock, ubxreader), daemon=True)
-    thr.start()
-    return thr
-
-
-def send_message(stream, lock, message):
-    """
-    Send message to device
+    Read queue and send UBX message to device
     """
 
-    lock.acquire()
-    stream.write(message.serialize())
-    lock.release()
+    while not stop.is_set():
+        if queue.empty() is False:
+            message = queue.get()
+            lock.acquire()
+            stream.write(message.serialize())
+            lock.release()
+            queue.task_done()
+
+
+def display_data(queue: Queue, stop: Event):
+    """
+    Get UBX data from queue and display.
+    """
+    # pylint: disable=unused-variable,
+
+    while not stop.is_set():
+        if queue.empty() is False:
+            (raw, parsed) = queue.get()
+            print(parsed)
+            queue.task_done()
 
 
 if __name__ == "__main__":
-
     # set port, baudrate and timeout to suit your device configuration
     if platform == "win32":  # Windows
         port = "COM13"
     elif platform == "darwin":  # MacOS
-        port = "/dev/tty.usbmodem14101"
+        port = "/dev/tty.usbmodem101"
     else:  # Linux
         port = "/dev/ttyACM1"
     baudrate = 9600
     timeout = 0.1
 
-    with Serial(port, baudrate, timeout=timeout) as serial:
+    with Serial(port, baudrate, timeout=timeout) as serial_stream:
+        ubxreader = UBXReader(serial_stream, protfilter=UBX_PROTOCOL)
 
-        # create UBXReader instance, reading only UBX messages
-        ubr = UBXReader(BufferedReader(serial), protfilter=2)
-
-        print("\nStarting read thread...\n")
-        reading = True
         serial_lock = Lock()
-        read_thread = start_thread(serial, serial_lock, ubr)
+        read_queue = Queue()
+        send_queue = Queue()
+        stop_event = Event()
 
-        # poll the receiver port configuration using CFG-PRT
-        print("\nPolling port configuration CFG-PRT...\n")
-        for prt in (0, 1, 2, 3, 4):  # I2C, UART1, UART2, USB, SPI
-            msg = UBXMessage("CFG", "CFG-PRT", POLL, portID=prt)
-            send_message(serial, serial_lock, msg)
-            sleep(1)
+        read_thread = Thread(
+            target=read_data,
+            args=(
+                serial_stream,
+                ubxreader,
+                read_queue,
+                serial_lock,
+                stop_event,
+            ),
+        )
+        write_thread = Thread(
+            target=write_data,
+            args=(
+                serial_stream,
+                send_queue,
+                serial_lock,
+                stop_event,
+            ),
+        )
+        display_thread = Thread(
+            target=display_data,
+            args=(
+                read_queue,
+                stop_event,
+            ),
+        )
 
-        # poll all available CFG configuration messages
-        print("\nPolling CFG configuration CFG-*...\n")
-        for (msgid, msgname) in UBX_MSGIDS.items():
-            if msgid[0] == 0x06:  # CFG-* configuration messages
-                msg = UBXMessage("CFG", msgname, POLL)
-                send_message(serial, serial_lock, msg)
+        print("\nStarting handler processes. Press Ctrl-C to terminate...")
+        read_thread.start()
+        write_thread.start()
+        display_thread.start()
+
+        # loop until user presses Ctrl-C
+        while not stop_event.is_set():
+            try:
+                # poll the receiver port configuration using CFG-PRT
+                print(f"\nPolling port configuration CFG-PRT...\n")
+                for prt in (0, 1, 2, 3, 4):  # I2C, UART1, UART2, USB, SPI
+                    msg = UBXMessage("CFG", "CFG-PRT", POLL, portID=prt)
+                    send_queue.put(msg)
                 sleep(1)
 
-        # poll a selection of current navigation message rates using CFG-MSG
-        print("\nPolling navigation message rates CFG-MSG...\n")
-        for (msgid, msgname) in UBX_MSGIDS.items():
-            if msgid[0] in (0x01, 0xF0, 0xF1):  # NAV, NMEA-Standard, NMEA-Proprietary
-                msg = UBXMessage("CFG", "CFG-MSG", POLL, payload=msgid)
-                send_message(serial, serial_lock, msg)
-                sleep(1)
+            except KeyboardInterrupt:  # capture Ctrl-C
+                print("\n\nTerminated by user.")
+                stop_event.set()
 
-        print("\nPolling complete. Pausing for any final responses...\n")
-        sleep(1)
-        print("\nStopping reader thread...\n")
-        reading = False
+        print("\nStop signal set. Waiting for threads to complete...")
         read_thread.join()
-        print("\nProcessing Complete")
+        write_thread.join()
+        display_thread.join()
+        print("\nProcessing complete")
